requirements.txt
django==6.0
djangorestframework==3.16.1

